Rename climaf html package into chtml to avoid conflict with python standard html packages.
